+++ plugins/editor/qml/Editor.qml
@@ -339,7 +339,6 @@ Rectangle{
                 readOnly: root.document === null || root.document.isMonitored || (importsShaped && rootShaped)
                 cursorVisible: !readOnly
 
-
                 Keys.onPressed: {
                     if ( event.key === Qt.Key_PageUp ){
                         if ( documentHandler.completionModel.isEnabled ){