+++ plugins/editor/qml/WorkspaceTheme.qml
@@ -1,10 +1,35 @@
 import QtQuick 2.3
 import editor 1.0
+import workspace 1.0
 
 Theme{
-    property color background: "#fff"
+    id: root
 
-    property color topBackground : "transparent"
+    // Color Palette
+
+    property QtObject colorScheme: QtObject{
+
+        property color background: '#03090d'
+        property color backgroundOverlay: '#060e13'
+
+        property color backgroundBorder: '#232f37'
+
+        property color middleground: '#141b20'
+        property color middlegroundOverlay: '#292d34'
+        property color middlegroundOverlayDominant: '#3f444d'
+
+        property color middlegroundBorder: '#232b30'
+        property color middlegroundHighlight: '#fff'
+
+        property color middlegroundOverlayDominantBorder: '#575b63'
+
+        property color foregroundFaded: '#dbdede'
+        property color foreground: '#fff'
+
+        property color textSelection: '#0b273f'
+    }
+
+    // Icons
 
     property string topNewIcon: "qrc:/images/top-icon-new.png"
     property string topSaveIcon: "qrc:/images/top-icon-save.png"
@@ -19,11 +44,204 @@ Theme{
     property string topOnSaveModeIcon : "qrc:/images/top-icon-mode-onsave.png"
     property string topDisabledModeIcon: "qrc:/images/top-icon-mode-disabled.png"
 
-    property color paneBackground: "#03090d"
-    property color paneTopBackground: "#0b1015"
-    property color paneTopBackgroundAlternate: "#050b12"
-    property color paneSplitterColor : "#10171f"
+    // Panes
+
+    property QtObject paneMenuStyle: QtObject{
+        property color backgroundColor: colorScheme.middleground
+        property double radius: 0
+
+        property QtObject itemStyle: TextStyle{
+            color: colorScheme.foregroundFaded
+            font : Qt.font({
+                family: 'Open Sans, sans-serif',
+                weight: Font.Light,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+        property QtObject itemHoverStyle: TextStyle{
+            color: Qt.lighter(colorScheme.foregroundFaded)
+            font : Qt.font({
+                family: 'Open Sans, sans-serif',
+                weight: Font.Light,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+    }
+
+    property color paneBackground: colorScheme.background
+    property color panebackgroundOverlay: colorScheme.backgroundOverlay
+    property color paneTopBackground: colorScheme.middleground
+    property color paneTopbackgroundOverlay: colorScheme.middleground
+    property color paneSplitterColor : colorScheme.middleground
 
     property color projectPaneItemBackground: "transparent"
     property color projectPaneItemEditBackground : "#1b2934"
+
+    // Forms
+
+    property color inputColor: '#3f444d'
+    property color inputBorderColor: '#fff'
+    property int   inputBorderWidth: 1
+
+    property QtObject inputLabelStyle: QtObject{
+        property color background: colorScheme.middleground
+        property double radius: 3
+        property QtObject textStyle: TextStyle{
+            color: colorScheme.foregroundFaded
+            font : Qt.font({
+                family: 'Open Sans, sans-serif',
+                weight: Font.Light,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+    }
+
+    property QtObject inputStyle: QtObject{
+        property QtObject textStyle: TextStyle{
+            color: colorScheme.foreground
+            font : Qt.font({
+                family: 'Open Sans, sans-serif',
+                weight: Font.Normal,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+        property double radius: 2
+        property QtObject hintTextStyle: TextStyle{
+            color: colorScheme.foreground
+            font : Qt.font({
+                family: 'Open Sans, sans-serif',
+                weight: Font.Normal,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+        property color backgroundColor: colorScheme.middleground
+        property color borderColor: colorScheme.backgroundBorder
+        property double borderThickness: 1
+        property color textSelectionColor: colorScheme.textSelection
+    }
+
+
+    property QtObject monoInputStyle: QtObject{
+        property QtObject textStyle: TextStyle{
+            color: colorScheme.foreground
+            font : Qt.font({
+                family: 'Source Code Pro, Ubuntu Mono, Courier New, Courier',
+                weight: Font.Normal,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+        property double radius: 3
+        property QtObject hintTextStyle: TextStyle{
+            color: 'white'
+            font : Qt.font({
+                family: 'Source Code Pro, Ubuntu Mono, Courier New, Courier',
+                weight: Font.Normal,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+        property color backgroundColor: colorScheme.background
+        property color borderColor: colorScheme.backgroundBorder
+        property double borderThickness: 1
+        property color textSelectionColor: colorScheme.textSelection
+    }
+
+
+    property QtObject formButtonStyle : QtObject{
+        property QtObject textStyle: TextStyle{
+            color: colorScheme.foreground
+            font : Qt.font({
+                family: 'Open Sans, sans-serif',
+                weight: Font.Normal,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+        property QtObject hoverTextStyle: TextStyle{
+            color: colorScheme.foreground
+            font : Qt.font({
+                family: 'Open Sans, sans-serif',
+                weight: Font.Normal,
+                italic: false,
+                pixelSize: 11
+            })
+        }
+        property color backgroundColor: colorScheme.middlegroundOverlayDominant
+        property color backgroundHoverColor: Qt.lighter(colorScheme.middlegroundOverlayDominant, 1.2)
+        property color borderColor: colorScheme.middlegroundOverlayDominantBorder
+        property color borderHoverColor: colorScheme.middlegroundOverlayDominantBorder
+        property double borderThickness: 1
+        property double radius: 3
+    }
+
+    // Buttons
+
+    property QtObject buttons: QtObject{
+
+        property Component apply: RectangleButton{
+            width: parent ? parent.width : 20
+            height: parent ? parent.height: 20
+
+            style: root.formButtonStyle
+
+            content: CheckMarkIcon{
+                anchors.centerIn: parent
+                width: parent.parent.width / 3
+                height: parent.parent.height / 3
+                strokeWidth: 1
+            }
+        }
+
+        property Component connect: RectangleButton{
+            width: parent ? parent.width : 20
+            height: parent ? parent.height: 20
+
+            style: root.formButtonStyle
+
+            content: Image{
+                anchors.centerIn: parent
+                source: "qrc:/images/palette-connections.png"
+            }
+        }
+    }
+
+    // Node Editor
+
+    property QtObject nodeEditor : QtObject{
+        property color backgroundColor: root.colorScheme.background
+        property color backgroundGridColor: root.colorScheme.middleground
+
+        property color borderColor: root.colorScheme.middlegroundOverlayDominant
+        property double borderWidth: 1
+        property double radius: 3
+
+        property color connectorEdgeColor: root.colorScheme.middlegroundOverlayDominant
+        property color connectorColor: root.colorScheme.middlegroundOverlayDominant
+
+        property color selectionColor: root.colorScheme.foreground
+        property double selectionWeight: 1
+
+        property QtObject objectNodeStyle : QtObject{
+            property color background: root.colorScheme.middleground
+            property double radius: 15
+            property color borderColor: root.colorScheme.middlegroundOverlayDominant
+            property double borderWidth: 1
+
+            property color titleBackground: root.colorScheme.middlegroundOverlayDominant
+            property double titleRadius: 3
+            property QtObject titleTextStyle : TextStyle{}
+        }
+
+        property QtObject propertyDelegateStyle : QtObject{
+            property color background: root.colorScheme.middlegroundOverlay
+            property double radius: 5
+            property QtObject textStyle: TextStyle{}
+        }
+    }
 }