+++ plugins/editqml/qml/ObjectContainer.qml
@@ -120,8 +120,8 @@ Item{
 
             if ( palette.item ){
                 var newPaletteBox = objectContainer.paletteContainerFactory.createObject(root.paletteGroup)
-                palette.item.x = 5
-                palette.item.y = 7
+                palette.item.x = 2
+                palette.item.y = 2
 
                 newPaletteBox.child = palette.item
                 newPaletteBox.palette = palette
@@ -154,7 +154,6 @@ Item{
         property QtObject editingFragment : null
         property Item editor: null
 
-
         property Item pane : null
         property Item wrapper : root
 
@@ -169,6 +168,8 @@ Item{
         property int topSpacing: 0
 
         property var propertiesOpened: []
+
+        property PaletteStyle paletteStyle: lk.layers.workspace.extensions.editqml.paletteStyle
 
         width: container.width < 260 ? 300 : container.width
         height: container.height < 10 || compact ? 40 : objectContainerTitleWrap.height + topSpacing + /*(paletteGroup ? paletteGroup.height : 0) +*/ container.height
@@ -246,7 +247,7 @@ Item{
                 ef.visualParent = paletteBoxGroup
 
                 childObjectContainer.paletteGroup = paletteBoxGroup
-                paletteBoxGroup.x = 5
+                paletteBoxGroup.x = 2
 
                 propertyContainer.valueContainer = childObjectContainer
                 childObjectContainer.expandDefaultPalette()
@@ -319,7 +320,7 @@ Item{
                             ef.visualParent = paletteBoxGroup
 
                             childObjectContainer.paletteGroup = paletteBoxGroup
-                            paletteBoxGroup.x = 5
+                            paletteBoxGroup.x = 2
 
                             propertyContainer.valueContainer = childObjectContainer
                             if ( propPalette ){
@@ -429,7 +430,6 @@ Item{
                 if (compact) expand()
                 else objectContainer.addPropertyFragmentToContainer(ef)
                 container.sortChildren()
-
             }
         }
 
@@ -443,7 +443,7 @@ Item{
                 id: objectContainerTitle
                 anchors.fill: parent
                 compact: objectContainer.compact
-                color: objectContainer.pane ? objectContainer.pane.topColor : '#062945'
+                color: objectContainer.pane ? objectContainer.pane.topColor : objectContainer.paletteStyle.sectionHeaderBackgroundColor//'#062945'
                 isBuilder: root.editingFragment ? root.editingFragment.isBuilder() : false
 
                 onToggleCompact: {
@@ -536,8 +536,8 @@ Item{
 
                             if ( palette.item ){
                                 var newPaletteBox = objectContainer.paletteContainerFactory.createObject(paletteGroup)
-                                palette.item.x = 5
-                                palette.item.y = 7
+                                palette.item.x = 2
+                                palette.item.y = 2
 
                                 newPaletteBox.child = palette.item
                                 newPaletteBox.palette = palette
@@ -736,7 +736,7 @@ Item{
             anchors.top: parent.top
             anchors.topMargin: objectContainerTitleWrap.height + topSpacing
             visible: !compact
-            spacing: 10
+            spacing: 5
             width: parentObjectContainer ? parentObjectContainer.width - (isForProperty? 140 : 20) : contentWidth + 10
 
             onChildrenChanged: recalculateContentWidth()
@@ -778,7 +778,7 @@ Item{
                 {
                     if (!children[i]) continue
                     if (children[i].objectName === "objectContainer"){
-                        children[i].topSpacing = 15
+                        children[i].topSpacing = 5
                         children[i].z = children.length - childrenCopy.length
                         childrenCopy.push(children[i])
                     }