+++ plugins/editqml/qml/PaletteControls.qml
@@ -1,0 +1,38 @@
+import QtQuick 2.0
+import editqml 1.0
+import editor 1.0
+
+QtObject{
+
+    property QtObject factories : QtObject{
+
+        property Component addQmlBox: Component{ AddQmlBox{} }
+        property Component paletteGroup: Component{ PaletteGroup{} }
+        property Component objectContainer: Component{ ObjectContainer{} }
+        property Component paletteContainer: Component{ PaletteContainer{} }
+        property Component paletteListView: Component{ PaletteListView{} }
+
+    }
+
+    function createAddQmlBox(parent){
+        return factories.addQmlBox.createObject(parent)
+    }
+
+    function createPaletteGroup(parent){
+        return factories.paletteGroup.createObject(parent)
+    }
+
+    function createObjectContainer(parent){
+        return factories.objectContainer.createObject(parent)
+    }
+
+    function createPaletteContainer(parent){
+        return factories.paletteContainer.createObject(parent)
+    }
+
+    function createPaletteListView(parent){
+        return factories.paletteListView.createObject(parent)
+    }
+
+
+}